requirements.txt
alembic==1.12.0
anyio==3.7.0
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.96.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
orjson==3.9.1
passlib==1.7.4
protobuf==3.20.3
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.9
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
